+++ lib/src/material/typeahead_field.dart
@@ -25,6 +25,7 @@ class TypeAheadField<T> extends RawTypeAheadField<T> {
     super.hideOnEmpty,
     super.hideOnError,
     super.hideOnLoading,
+    super.showOnFocus,
     super.hideOnUnfocus,
     super.hideWithKeyboard,
     super.hideOnSelect,